api/requirements.txt
fastapi~=0.100
pymongo~=4.4
uvicorn~=0.23
python-dotenv~=1.0
python-multipart~=0.0